requirements.txt
ddgs>=1.0.0
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
anthropic>=0.25.0
fpdf2>=2.7.9
python-dotenv>=1.0.0
gunicorn>=21.2.0
supabase>=2.0.0
